move nanovg render to module
