+++ hvcc/generators/c2fmod/c2fmod.py
@@ -26,7 +26,7 @@ class c2fmod(Generator):
         c_src_dir: str,
         out_dir: str,
         externs: ExternInfo,
-        patch_name: Optional[str] = None,
+        patch_name: str,
         patch_meta: Meta = Meta(),
         num_input_channels: int = 0,
         num_output_channels: int = 0,
