+++ hvcc/generators/c2wwise/c2wwise.py
@@ -35,16 +35,16 @@ class c2wwise(Generator):
 
     @classmethod
     def compile(
-            cls,
-            c_src_dir: str,
-            out_dir: str,
-            externs: ExternInfo,
-            patch_name: Optional[str] = None,
-            patch_meta: Meta = Meta(),
-            num_input_channels: int = 0,
-            num_output_channels: int = 0,
-            copyright: Optional[str] = None,
-            verbose: Optional[bool] = False
+        cls,
+        c_src_dir: str,
+        out_dir: str,
+        externs: ExternInfo,
+        patch_name: str,
+        patch_meta: Meta = Meta(),
+        num_input_channels: int = 0,
+        num_output_channels: int = 0,
+        copyright: Optional[str] = None,
+        verbose: Optional[bool] = False
     ) -> CompilerResp:
         tick = time.time()
