+++ hvcc/types/compiler.py
@@ -81,7 +81,7 @@ class Generator(ABC):
         c_src_dir: str,
         out_dir: str,
         externs: ExternInfo,
-        patch_name: Optional[str] = None,
+        patch_name: str,
         patch_meta: Meta = Meta(),
         num_input_channels: int = 0,
         num_output_channels: int = 0,
